Show 2 datepickers
